+++ flask_monitoringdashboard/__init__.py
@@ -17,9 +17,9 @@ import os
 from flask import Blueprint
 
 from flask_monitoringdashboard.core.config import Config
+from flask_monitoringdashboard.core.logger import log
 
 config = Config()
-user_app = None
 
 
 # get current location of the project
@@ -31,43 +31,42 @@ def loc():
 blueprint = Blueprint('dashboard', __name__, template_folder=loc() + 'templates')
 
 
-def bind(app):
+def bind(app, schedule=True):
     """
         Binding the app to this object should happen before importing the routing-
         methods below. Thus, the importing statement is part of this function.
         :param app: the app for which the performance has to be tracked
+        :param schedule: flag telling if the background scheduler should be started
     """
-    assert app is not None
-    global user_app, blueprint
-    user_app = app
-
+    config.app = app
     # Provide a secret-key for using WTF-forms
-    if not user_app.secret_key:
-        print('WARNING: You should provide a security key.')
-        user_app.secret_key = 'my-secret-key'
-
-    import os
-    # Only initialize unit test logging when running on Travis.
-    if '/home/travis/build/' in os.getcwd():
-        print('Detected running on Travis.')
-        import datetime
-        from flask import request
-
-        @user_app.after_request
-        def after_request(response):
-            hit_time_stamp = str(datetime.datetime.now())
-            home = os.path.expanduser("~")
-            log = open(home + '/endpoint_hits.log', 'a')
-            log.write('"{}","{}"\n'.format(hit_time_stamp, request.endpoint))
-            log.close()
-            return response
+    if not app.secret_key:
+        log('WARNING: You should provide a security key.')
+        app.secret_key = 'my-secret-key'
 
     # Add all route-functions to the blueprint
+    from flask_monitoringdashboard.views import deployment, custom, endpoint, outlier, request, profiler, version, auth
     import flask_monitoringdashboard.views
 
     # Add wrappers to the endpoints that have to be monitored
     from flask_monitoringdashboard.core.measurement import init_measurement
+    from flask_monitoringdashboard.core import custom_graph
+
     blueprint.before_app_first_request(init_measurement)
+    if schedule:
+        custom_graph.init(app)
 
     # register the blueprint to the app
     app.register_blueprint(blueprint, url_prefix='/' + config.link)
+
+
+def add_graph(title, func, **schedule):
+    """
+    Add a custom graph to the dashboard. You must specify the following arguments
+    :param title: title of the graph (must be unique)
+    :param schedule: dict containing values for weeks, days, hours, minutes, seconds
+    :param func: function reference without arguments
+    """
+    from flask_monitoringdashboard.core import custom_graph
+    graph_id = custom_graph.register_graph(title)
+    custom_graph.add_background_job(func, graph_id, **schedule)