+++ flask_monitoringdashboard/core/auth.py
@@ -1,14 +1,16 @@
 from functools import wraps
-from flask_monitoringdashboard import config
+
 from flask import session, redirect, url_for
+
+from flask_monitoringdashboard import config
 
 
 def admin_secure(func):
     """
-        When the user is not logged into the system, the user is requested to the login page.
-        There are two types of user-modes:
-        - admin: Can be visited with this wrapper.
-        - guest: Cannot be visited with this wrapper.
+    When the user is not logged into the system, the user is requested to the login page.
+    There are two types of user-modes:
+    - admin: Can be visited with this wrapper.
+    - guest: Cannot be visited with this wrapper.
     :param func: the endpoint to be wrapped.
     """
 
@@ -24,10 +26,10 @@ def admin_secure(func):
 
 def secure(func):
     """
-        When the user is not logged into the system, the user is requested to the login page.
-        There are two types of user-modes:
-        - admin: Can be visited with this wrapper.
-        - guest: Can be visited with this wrapper.
+    When the user is not logged into the system, the user is requested to the login page.
+    There are two types of user-modes:
+    - admin: Can be visited with this wrapper.
+    - guest: Can be visited with this wrapper.
     :param func: the endpoint to be wrapped.
     """