+++ flask_monitoringdashboard/core/group_by.py
@@ -1,14 +1,17 @@
 from flask_monitoringdashboard import config
+from flask_monitoringdashboard.core.logger import log
 
 PRIMITIVES = (bool, bytes, float, int, str)
 
 
 def recursive_group_by(argument):
-    """ Returns the result of the given argument. The result is computed as:
+    """
+    Returns the result of the given argument. The result is computed as:
     - If the argument is a primitive (i.e. str, bool, int, ...) return its value.
     - If the argument is a function, call the function.
     - If the argument is iterable (i.e. list or tuple), compute the result by iterating over the argument
-    Return type is always a string"""
+    Return type is always a string
+    """
 
     if type(argument) in PRIMITIVES:
         return str(argument)
@@ -35,5 +38,5 @@ def get_group_by():
         if config.group_by:
             group_by = recursive_group_by(config.group_by)
     except Exception as e:
-        print('Can\'t execute group_by function: {}'.format(e))
+        log('Can\'t execute group_by function: {}'.format(e))
     return str(group_by)