+++ flask_monitoringdashboard/core/utils.py
@@ -1,50 +1,59 @@
-import ast
-
 import numpy as np
 from flask import url_for
 from werkzeug.routing import BuildError
 
 from flask_monitoringdashboard import config
-from flask_monitoringdashboard.database.count import count_requests
-from flask_monitoringdashboard.database.endpoint import get_monitor_rule
-from flask_monitoringdashboard.database.function_calls import get_date_of_first_request
+from flask_monitoringdashboard.core.colors import get_color
+from flask_monitoringdashboard.core.rules import get_rules
+from flask_monitoringdashboard.core.timezone import to_local_datetime
+from flask_monitoringdashboard.database.count import count_requests, count_total_requests
+from flask_monitoringdashboard.database.endpoint import get_endpoint_by_id
+from flask_monitoringdashboard.database.request import get_date_of_first_request, get_date_of_first_request_version
 
 
-def get_endpoint_details(db_session, endpoint):
-    """ Return details about an endpoint"""
+def get_endpoint_details(db_session, endpoint_id):
+    """
+    Returns details about an endpoint.
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint
+    :return dictionary
+    """
+    endpoint = get_endpoint_by_id(db_session, endpoint_id)
+    endpoint.time_added = to_local_datetime(endpoint.time_added)
+    flask_rule = get_rules(endpoint.name)
+    methods = [list(rule.methods) for rule in flask_rule]
+    methods = sum(methods, [])  # flatten list
     return {
-        'endpoint': endpoint,
-        'rule': get_monitor_rule(db_session, endpoint),
-        'url': get_url(endpoint),
-        'total_hits': count_requests(db_session, endpoint)
+        'id': endpoint_id,
+        'color': get_color(endpoint.name),
+        'methods': list(dict.fromkeys(methods)),
+        'endpoint': endpoint.name,
+        'rules': [r.rule for r in get_rules(endpoint.name)],
+        'monitor-level': endpoint.monitor_level,
+        'url': get_url(endpoint.name),
+        'total_hits': count_requests(db_session, endpoint.id)
     }
 
 
 def get_details(db_session):
-    """ Return details about the deployment """
+    """
+    Returns details about the deployment.
+    :param db_session: session for the database
+    :return dictionary
+    """
     import json
-    with open('flask_monitoringdashboard/constants.json', 'r') as f:
+    from flask_monitoringdashboard import loc
+    with open(loc() + 'constants.json', 'r') as f:
         constants = json.load(f)
 
     return {
         'link': config.link,
         'dashboard-version': constants['version'],
         'config-version': config.version,
-        'first-request': get_date_of_first_request(db_session)
+        'first-request': get_date_of_first_request(db_session),
+        'first-request-version': get_date_of_first_request_version(db_session, config.version),
+        'total-requests': count_total_requests(db_session)
     }
-
-
-def formatter(ms):
-    """
-    formats the ms into seconds and ms
-    :param ms: the number of ms
-    :return: a string representing the same amount, but now represented in seconds and ms.
-    """
-    sec = int(ms) // 1000
-    ms = int(ms) % 1000
-    if sec == 0:
-        return '{0}ms'.format(ms)
-    return '{0}.{1}s'.format(sec, ms)
 
 
 def get_url(end):
@@ -67,27 +76,6 @@ def simplify(values, n=5):
     :param n: length of the returned list
     :return: list with n values: min, q1, median, q3, max
     """
+    if len(values) <= n:
+        return values
     return [np.percentile(values, i * 100 // (n - 1)) for i in range(n)]
-
-
-def get_mean_cpu(cpu_percentages):
-    """
-    Returns a list containing mean CPU percentages per core for all given CPU percentages.
-    :param cpu_percentages: list of CPU percentages
-    """
-    if not cpu_percentages:
-        return None
-
-    count = 0  # some outliers have no CPU info
-    values = []  # list of lists that stores the CPU info
-
-    for cpu in cpu_percentages:
-        if not cpu:
-            continue
-        x = ast.literal_eval(cpu[0])
-        values.append(x)
-        count += 1
-
-    sums = [sum(x) for x in zip(*values)]
-    means = list(map(lambda x: round(x / count), sums))
-    return means