+++ flask_monitoringdashboard/database/count.py
@@ -1,11 +1,12 @@
 from sqlalchemy import func, distinct
 
-from flask_monitoringdashboard.database import FunctionCall, Outlier
+from flask_monitoringdashboard.database import Request, StackLine
 
 
 def count_rows(db_session, column, *criterion):
     """
-    Count the number of rows of a specified column
+    Count the number of rows of a specified column.
+    :param db_session: session for the database
     :param column: column to count
     :param criterion: where-clause of the query
     :return: number of rows
@@ -16,40 +17,44 @@ def count_rows(db_session, column, *criterion):
     return 0
 
 
-def count_users(db_session, endpoint):
+def count_requests(db_session, endpoint_id, *where):
     """
-    :param endpoint: filter on this endpoint
-    :return: The number of distinct users that have requested this endpoint
-    """
-    return count_rows(db_session, FunctionCall.group_by, FunctionCall.endpoint == endpoint)
-
-
-def count_ip(db_session, endpoint):
-    """
-    :param endpoint: filter on this endpoint
-    :return: The number of distinct users that have requested this endpoint
-    """
-    return count_rows(db_session, FunctionCall.ip, FunctionCall.endpoint == endpoint)
-
-
-def count_versions(db_session, endpoint):
-    """
-    :param endpoint: filter on this endpoint
-    :return: The number of distinct versions that are used for this endpoint
-    """
-    return count_rows(db_session, FunctionCall.version, FunctionCall.endpoint == endpoint)
-
-
-def count_requests(db_session, endpoint, *where):
-    """ Return the number of hits for a specific endpoint (possible with more filter arguments).
-    :param endpoint: name of the endpoint
+    Return the number of hits for a specific endpoint (possible with more filter arguments).
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint
     :param where: additional arguments
     """
-    return count_rows(db_session, FunctionCall.id, FunctionCall.endpoint == endpoint, *where)
+    return count_rows(db_session, Request.id, Request.endpoint_id == endpoint_id, *where)
 
 
-def count_outliers(db_session, endpoint):
+def count_total_requests(db_session, *where):
     """
+    Return the number of total hits
+    :param db_session: session for the database
+    :param where: additional arguments
+    """
+    return count_rows(db_session, Request.id, *where)
+
+
+def count_outliers(db_session, endpoint_id):
+    """
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint
     :return: An integer with the number of rows in the Outlier-table.
     """
-    return count_rows(db_session, Outlier.id, Outlier.endpoint == endpoint)
+    return count_rows(db_session, Request.id, Request.endpoint_id == endpoint_id, Request.outlier)
+
+
+def count_profiled_requests(db_session, endpoint_id):
+    """
+    Count the number of profiled requests for a certain endpoint
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint
+    :return: An integer
+    """
+    count = db_session.query(func.count(distinct(StackLine.request_id))). \
+        filter(Request.endpoint_id == endpoint_id). \
+        join(Request.stack_lines).first()
+    if count:
+        return count[0]
+    return 0