+++ flask_monitoringdashboard/database/outlier.py
@@ -1,48 +1,54 @@
-import datetime
-
-from flask import json
 from sqlalchemy import desc
+from sqlalchemy.orm import joinedload
 
-from flask_monitoringdashboard.database import Outlier
+from flask_monitoringdashboard.database import Outlier, Request
 
 
-def add_outlier(db_session, endpoint, execution_time, stack_info, request):
-    """ Collects information (request-parameters, memory, stacktrace) about the request and adds it in the database."""
-    outlier = Outlier(endpoint=endpoint, request_values=json.dumps(request.values),
-                      request_headers=str(request.headers), request_environment=str(request.environ),
-                      request_url=str(request.url), cpu_percent=stack_info.cpu_percent,
-                      memory=stack_info.memory, stacktrace=stack_info.stacktrace,
-                      execution_time=execution_time, time=datetime.datetime.now())
+def add_outlier(db_session, request_id, cpu_percent, memory, stacktrace, request):
+    """
+    Adds an Outlier object in the database.
+    :param db_session: session for the database
+    :param request_id: id of the request
+    :param cpu_percent: cpu load of the server when processing the request
+    :param memory: memory load of the server when processing the request
+    :param stacktrace: stack trace of the request
+    :param request: triple containing the headers, environment and url
+    """
+    headers, environ, url = request
+    outlier = Outlier(request_id=request_id, request_header=headers,
+                      request_environment=environ,
+                      request_url=url, cpu_percent=cpu_percent,
+                      memory=memory, stacktrace=stacktrace)
     db_session.add(outlier)
 
 
-def get_outliers_sorted(db_session, endpoint, sort_column, offset, per_page):
+def get_outliers_sorted(db_session, endpoint_id, offset, per_page):
     """
-    :param endpoint: only get outliers from this endpoint
-    :param sort_column: column used for sorting the result
+    Gets a list of Outlier objects for a certain endpoint, sorted by most recent request time
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint for filtering the requests
     :param offset: number of items to skip
     :param per_page: number of items to return
-    :return: a list of all outliers of a specific endpoint. The list is sorted based on the column that is given.
+    :return list of Outlier objects of a specific endpoint
     """
-    result = db_session.query(Outlier).filter(Outlier.endpoint == endpoint).order_by(desc(sort_column)). \
+    result = db_session.query(Outlier). \
+        join(Outlier.request). \
+        options(joinedload(Outlier.request).joinedload(Request.endpoint)). \
+        filter(Request.endpoint_id == endpoint_id). \
+        order_by(desc(Request.time_requested)). \
         offset(offset).limit(per_page).all()
     db_session.expunge_all()
     return result
 
 
-def get_outliers_cpus(db_session, endpoint):
+def get_outliers_cpus(db_session, endpoint_id):
     """
-    :param db_session: the session containing the query
-    :param endpoint: only get outliers from this endpoint
-    :return: a list of all cpu percentages for outliers of a specific endpoint
+    Gets list of CPU loads of all outliers of a certain endpoint
+    :param db_session: session for the database
+    :param endpoint_id: id of the endpoint
+    :return list of cpu percentages as strings
     """
-    result = db_session.query(Outlier.cpu_percent).filter(Outlier.endpoint == endpoint).all()
-    return result
-
-
-def delete_outliers_without_stacktrace(db_session):
-    """
-        Remove the outliers which don't have a stacktrace.
-        This is possibly due to an error in the outlier functionality
-    """
-    db_session.query(Outlier).filter(Outlier.stacktrace == '').delete()
+    outliers = db_session.query(Outlier.cpu_percent). \
+        join(Outlier.request). \
+        filter(Request.endpoint_id == endpoint_id).all()
+    return [outlier[0] for outlier in outliers]