+++ flask_monitoringdashboard/test/db/test_endpoint.py
@@ -1,14 +1,14 @@
 """
     This file contains all unit tests for the endpoint-table in the database. (Corresponding to the file:
     'flask_monitoringdashboard/database/endpoint.py')
-    See __init__.py for how to run the test-cases.
+    See info_box.py for how to run the test-cases.
 """
 
 import unittest
 
+from flask_monitoringdashboard.core.timezone import to_utc_datetime
 from flask_monitoringdashboard.database import session_scope
-from flask_monitoringdashboard.test.utils import set_test_environment, clear_db, add_fake_data, EXECUTION_TIMES, NAME, \
-    GROUP_BY, IP
+from flask_monitoringdashboard.test.utils import set_test_environment, clear_db, add_fake_data, NAME, TIMES
 
 
 class TestEndpoint(unittest.TestCase):
@@ -18,54 +18,53 @@ class TestEndpoint(unittest.TestCase):
         clear_db()
         add_fake_data()
 
-    def test_get_monitor_rule(self):
+    def test_get_endpoint(self):
         """
             Test wheter the function returns the right values.
         """
-        from flask_monitoringdashboard.database.endpoint import get_monitor_rule
+        from flask_monitoringdashboard.database.endpoint import get_endpoint_by_name
         from flask_monitoringdashboard import config
         with session_scope() as db_session:
-            rule = get_monitor_rule(db_session, NAME)
-        self.assertEqual(rule.endpoint, NAME)
-        self.assertTrue(rule.monitor)
-        self.assertEqual(rule.version_added, config.version)
+            endpoint = get_endpoint_by_name(db_session, NAME)
+        self.assertEqual(endpoint.name, NAME)
+        self.assertEqual(endpoint.monitor_level, 1)
+        self.assertEqual(endpoint.version_added, config.version)
 
-    def test_update_monitor_rule(self):
+    def test_update_endpoint(self):
         """
             Test whether the function returns the right values.
         """
-        from flask_monitoringdashboard.database.endpoint import get_monitor_rule, update_monitor_rule
+        from flask_monitoringdashboard.database.endpoint import get_endpoint_by_name, update_endpoint
         with session_scope() as db_session:
-            current_value = get_monitor_rule(db_session, NAME).monitor
-            new_value = not current_value
-            update_monitor_rule(db_session, NAME, new_value)
-            self.assertEqual(get_monitor_rule(db_session, NAME).monitor, new_value)
-
-    def test_get_all_measurement_per_column(self):
-        """
-            Test whether the function returns the right values.
-        """
-        from flask_monitoringdashboard.database.endpoint import get_all_measurement_per_column
-        from flask_monitoringdashboard.database import FunctionCall
-        from flask_monitoringdashboard import config
-        with session_scope() as db_session:
-            result = get_all_measurement_per_column(db_session, endpoint=NAME, column=FunctionCall.ip, value=IP)
-            self.assertEqual(len(result), len(EXECUTION_TIMES))
-            for row in result:
-                self.assertIn(row.execution_time, EXECUTION_TIMES)
-                self.assertEqual(row.version, config.version)
-                self.assertEqual(row.endpoint, NAME)
-                self.assertEqual(row.group_by, GROUP_BY)
-                self.assertEqual(row.ip, IP)
+            current_value = get_endpoint_by_name(db_session, NAME).monitor_level
+            new_value = 1 if current_value != 1 else 2
+            update_endpoint(db_session, NAME, new_value)
+            self.assertEqual(get_endpoint_by_name(db_session, NAME).monitor_level, new_value)
 
     def test_update_last_accessed(self):
         """
             Test whether the function returns the right values.
         """
         import datetime
-        time = datetime.datetime.now()
-        from flask_monitoringdashboard.database.endpoint import update_last_accessed, get_last_accessed_times
+        time = datetime.datetime.utcnow()
+        from flask_monitoringdashboard.database.endpoint import update_last_accessed, get_last_requested
+        from flask_monitoringdashboard.database.count_group import get_value
         with session_scope() as db_session:
-            update_last_accessed(db_session, NAME, time)
-            result = get_last_accessed_times(db_session, NAME)
-            self.assertEqual(result, time)
+            update_last_accessed(db_session, NAME)
+            result = get_value(get_last_requested(db_session), NAME)
+            result_utc = to_utc_datetime(result)
+            self.assertTrue((result_utc - time).seconds < 1)
+
+    def test_endpoints(self):
+        """
+            Test whether the function returns the right values.
+        """
+        from flask_monitoringdashboard.database.endpoint import get_endpoints
+        from flask_monitoringdashboard import config
+        with session_scope() as db_session:
+            result = get_endpoints(db_session)
+            self.assertEqual(len(result), 1)
+            self.assertEqual(result[0].name, NAME)
+            self.assertEqual(result[0].monitor_level, 1)
+            self.assertEqual(result[0].version_added, config.version)
+            self.assertEqual(result[0].last_requested, TIMES[0])