+++ docs/changelog.rst
@@ -5,17 +5,82 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Please note that the changes before version 1.10.0 have not been documented.
 
-
-Unreleased
+v2.1.1
 ----------
 Changed
 
-- Added mean CPU
+- Default monitoring level is now 1
+- Fixed bug causing config file not being parsed
+- Monitoring level can be set from the 'detail' section
+- Improved README
 
+v2.1.0
+----------
+Changed
+
+- Frontend is now using AngularJS
+- Removed TestMonitor
+- Added Custom graphs
+- Fixed Issue #206
+- Added support for Python 3.7
 - Updated documentation
+- Updated unit tests
+
+v2.0.7
+----------
+Changed
+
+- Fixed Issue #174
+
+- Fixed issue with profiler not going into code
+
+- Implemented a Sunburst visualization of the Grouped Profiler
+
+- Improved test coverage
+
+- Improved python-doc
+
+- Added functionality to download the outlier data
+
+- Dropped support for Python 3.3 and 3.4
+
+
+v2.0.0
+----------
+Changed
+
+- Added a configuration option to prefix a table in the database
+
+- Optimize queries, such that viewing data is faster
+
+- Updated database scheme
+
+- Implemented functionality to customize time window of graphs
+
+- Implemented a profiler for Request profiling
+
+- Implemented a profiler for Endpoint profiling
+
+- Refactored current code, which improves readability
+
+- Refactoring of Test-Monitoring page
+
+- Identify testRun by Travis build number
+
+
+v1.13.0
+----------
+Changed
+
+- Added boxplot of CPU loads
+
+- Updated naming scheme of all graphs
+
+- Implemented two configuration options: the local timezone and the option to automatically monitor new endpoints
 
 - Updated the Test-Monitoring initialization
 
+- Updated Database support for MySQL
 
 v1.12.0
 -------